refactor(calculator) : 자동차 경주게임 기능정리
